# home/springdoc/openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"security":[{"bearerAuth":[]}],"paths":{"/v1/users":{"get":{"tags":["user-controller"],"operationId":"getUsers","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserDto"}}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["user-controller"],"operationId":"editUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["user-controller"],"operationId":"createUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/users/own":{"put":{"tags":["user-controller"],"operationId":"editUserUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/rooms":{"get":{"tags":["room-controller"],"operationId":"getRooms","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomDto"}}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["room-controller"],"operationId":"editRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["room-controller"],"operationId":"createRoom","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RoomWithBookingsDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/localizations":{"get":{"tags":["localization-controller"],"operationId":"getLocalizations","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocalizationDto"}}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["localization-controller"],"operationId":"editLocalization","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalizationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalizationDto"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["localization-controller"],"operationId":"createLocalization","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocalizationDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalizationDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/hotels":{"get":{"tags":["hotel-controller"],"operationId":"getHotels","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HotelDto"}}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["hotel-controller"],"operationId":"editHotel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["hotel-controller"],"operationId":"createHotel","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HotelWithRoomsDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/bookings":{"get":{"tags":["booking-controller"],"operationId":"getBookings","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookingDto"}}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["booking-controller"],"operationId":"editBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["booking-controller"],"operationId":"createBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/bookings/own":{"get":{"tags":["booking-controller"],"operationId":"getOwnBookings","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/BookingDto"}}}}}},"security":[{"bearerAuth":[]}]},"put":{"tags":["booking-controller"],"operationId":"editOwnBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["booking-controller"],"operationId":"createOwnBooking","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BookingDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/users/registration":{"post":{"tags":["user-controller"],"operationId":"createUserRegistration","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/initializeDb":{"post":{"tags":["manager-controller"],"operationId":"initializeDatabase","responses":{"200":{"description":"OK"}}}},"/login":{"post":{"tags":["login-controller"],"operationId":"LoginUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginCredentials"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/v1/users/own/bookings":{"get":{"tags":["user-controller"],"operationId":"getSingleUserOwner","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithBookingDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/users/bookings":{"get":{"tags":["user-controller"],"operationId":"getUsersWithBookings","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserWithBookingDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/v1/users/bookings/{id}":{"get":{"tags":["user-controller"],"operationId":"getSingleUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/UserWithBookingDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/rooms/bookings":{"get":{"tags":["room-controller"],"operationId":"getRoomsWithBookings","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomWithBookingsDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/v1/rooms/bookings/{id}":{"get":{"tags":["room-controller"],"operationId":"getSingleRoomwithBookings","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomWithBookingsDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/rooms/bookings/withoutUsers":{"get":{"tags":["room-controller"],"operationId":"getRoomsWithBookingsWithoutUsers","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/RoomWithBookingsWithoutUsersDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/v1/rooms/bookings/withoutUsers/{id}":{"get":{"tags":["room-controller"],"operationId":"getSingleRoomWithBookingsWithoutUsers","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/RoomWithBookingsWithoutUsersDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/localizations/hotels":{"get":{"tags":["localization-controller"],"operationId":"getLocalizationsWithHotels","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocalizationWithHotelsDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/v1/localizations/hotels/{id}":{"get":{"tags":["localization-controller"],"operationId":"getSingleLocalizationWithHotels","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocalizationWithHotelsDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/hotels/rooms":{"get":{"tags":["hotel-controller"],"operationId":"getHotelsWithRooms","parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","format":"int32"}},{"name":"sort","in":"query","required":true,"schema":{"type":"string","enum":["ASC","DESC"]}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HotelWithRoomsDto"}}}}}},"security":[{"bearerAuth":[]}]}},"/v1/hotels/rooms/{id}":{"get":{"tags":["hotel-controller"],"operationId":"getSingleHotelWithRooms","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/HotelWithRoomsDto"}}}}},"security":[{"bearerAuth":[]}]}},"/v1/bookings/{id}":{"get":{"tags":["booking-controller"],"operationId":"getSingleBooking","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/BookingDto"}}}}},"security":[{"bearerAuth":[]}]},"delete":{"tags":["booking-controller"],"operationId":"deleteBooking","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/v1/users/{id}":{"delete":{"tags":["user-controller"],"operationId":"deleteUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/v1/rooms/{id}":{"delete":{"tags":["room-controller"],"operationId":"deleteRoom","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/v1/localizations/{id}":{"delete":{"tags":["localization-controller"],"operationId":"deleteSingleLocalization","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/v1/hotels/{id}":{"delete":{"tags":["hotel-controller"],"operationId":"deleteHotel","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}},"/v1/clearDb":{"delete":{"tags":["manager-controller"],"operationId":"clearDatabase","responses":{"200":{"description":"OK"}}}},"/v1/bookings/own/{id}":{"delete":{"tags":["booking-controller"],"operationId":"deleteOwnBooking","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"UserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"email":{"type":"string"}}},"RoomDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roomNumber":{"type":"integer","format":"int32"},"numberOfPersons":{"type":"integer","format":"int32"},"standard":{"type":"integer","format":"int32"},"hotelId":{"type":"integer","format":"int64"}}},"LocalizationDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"city":{"type":"string"},"country":{"type":"string"}}},"HotelDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"numberOfStars":{"type":"integer","format":"int32"},"hotelChain":{"type":"string"},"localizationId":{"type":"integer","format":"int64"}}},"BookingDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"roomId":{"type":"integer","format":"int64"},"start_date":{"type":"string","format":"date"},"end_date":{"type":"string","format":"date"}}},"RoomWithBookingsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roomNumber":{"type":"integer","format":"int32"},"numberOfPersons":{"type":"integer","format":"int32"},"standard":{"type":"integer","format":"int32"},"hotelId":{"type":"integer","format":"int64"},"bookings":{"type":"array","items":{"$ref":"#/components/schemas/BookingDto"}}}},"HotelWithRoomsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"numberOfStars":{"type":"integer","format":"int32"},"hotelChain":{"type":"string"},"localizationId":{"type":"integer","format":"int64"},"rooms":{"type":"array","items":{"$ref":"#/components/schemas/RoomDto"}}}},"LoginCredentials":{"type":"object","properties":{"username":{"type":"string"},"password":{"type":"string"}}},"Booking":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"userId":{"type":"integer","format":"int64"},"roomId":{"type":"integer","format":"int64"},"start_date":{"type":"string","format":"date"},"end_date":{"type":"string","format":"date"}}},"UserWithBookingDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"firstName":{"type":"string"},"lastName":{"type":"string"},"dateOfBirth":{"type":"string","format":"date"},"username":{"type":"string"},"password":{"type":"string"},"role":{"type":"string"},"email":{"type":"string"},"bookings":{"type":"array","items":{"$ref":"#/components/schemas/Booking"}}}},"BookingWithoutUserDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roomId":{"type":"integer","format":"int64"},"start_date":{"type":"string","format":"date"},"end_date":{"type":"string","format":"date"}}},"RoomWithBookingsWithoutUsersDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"roomNumber":{"type":"integer","format":"int32"},"numberOfPersons":{"type":"integer","format":"int32"},"standard":{"type":"integer","format":"int32"},"hotelId":{"type":"integer","format":"int64"},"bookings":{"type":"array","items":{"$ref":"#/components/schemas/BookingWithoutUserDto"}}}},"LocalizationWithHotelsDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"city":{"type":"string"},"country":{"type":"string"},"hotel":{"type":"array","items":{"$ref":"#/components/schemas/HotelDto"}}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}